[k2] move kphp component entry point and state to components/kphp/
Prepare runtime-light for multiple component entry points: the kphp
component's sources (runtime-light.cpp and state/) move to
components/kphp/, making room for sibling components (e.g. confdata).
Includes are updated accordingly; .gitignore's bare "kphp" pattern is
anchored to the repo root so it no longer hides the new directory.
