+++ runtime-light/components/kphp/kphp.cmake
@@ -1,0 +1,2 @@
+prepend(RUNTIME_LIGHT_KPHP_COMPONENT_SRC components/kphp/ kphp-component.cpp
+        state/component-state.cpp state/instance-state.cpp)
