+++ runtime-light/server/http/multipart/details/parts-processing.cpp
@@ -20,10 +20,10 @@
 #include "runtime-common/core/runtime-core.h"
 #include "runtime-common/core/std/containers.h"
 #include "runtime-common/stdlib/server/url-functions.h"
+#include "runtime-light/components/kphp/state/component-state.h"
 #include "runtime-light/k2-platform/k2-api.h"
 #include "runtime-light/server/http/http-server-state.h"
 #include "runtime-light/server/http/multipart/details/parts-parsing.h"
-#include "runtime-light/state/component-state.h"
 #include "runtime-light/stdlib/diagnostics/logs.h"
 #include "runtime-light/stdlib/file/resource.h"
 #include "runtime-light/stdlib/math/random-functions.h"
