--- runtime-light/state/state.cmake
@@ -1,1 +1,0 @@
-prepend(RUNTIME_LIGHT_STATE_SRC state/ component-state.cpp instance-state.cpp)
