+++ runtime-light/stdlib/diagnostics/backtrace.cpp
@@ -6,10 +6,10 @@
 #include <iterator>
 #include <span>
 
+#include "runtime-light/components/kphp/state/instance-state.h"
 #include "runtime-light/coroutine/async-stack.h"
 #include "runtime-light/coroutine/coroutine-state.h"
 #include "runtime-light/k2-platform/k2-api.h"
-#include "runtime-light/state/instance-state.h"
 #include "runtime-light/stdlib/diagnostics/backtrace.h"
 #include "runtime-light/stdlib/diagnostics/logs.h"
