+++ runtime-light/stdlib/diagnostics/contextual-tags.cpp
@@ -8,8 +8,8 @@
 #include <functional>
 #include <optional>
 
+#include "runtime-light/components/kphp/state/instance-state.h"
 #include "runtime-light/k2-platform/k2-api.h"
-#include "runtime-light/state/instance-state.h"
 
 namespace kphp::log {
