+++ runtime-light/stdlib/job-worker/job-worker-api.cpp
@@ -11,11 +11,11 @@
 #include <utility>
 
 #include "runtime-common/core/runtime-core.h"
+#include "runtime-light/components/kphp/state/instance-state.h"
 #include "runtime-light/coroutine/io-scheduler.h"
 #include "runtime-light/coroutine/task.h"
 #include "runtime-light/k2-platform/k2-api.h"
 #include "runtime-light/server/job-worker/job-worker-server-state.h"
-#include "runtime-light/state/instance-state.h"
 #include "runtime-light/stdlib/component/component-api.h"
 #include "runtime-light/stdlib/fork/fork-functions.h"
 #include "runtime-light/stdlib/fork/fork-state.h"
