+++ runtime-light/stdlib/system/system-state.cpp
@@ -4,7 +4,7 @@
 
 #include "runtime-light/stdlib/system/system-state.h"
 
-#include "runtime-light/state/instance-state.h"
+#include "runtime-light/components/kphp/state/instance-state.h"
 
 SystemInstanceState& SystemInstanceState::get() noexcept {
   return InstanceState::get().system_instance_state;
